+++ laik_vector.h
@@ -116,90 +116,4 @@ protected:
     void init_config_params(Laik_Group* group);
 };
 
-template <typename T>
-class laik_vector_halo:public laik_vector<T>
-{
-public:
-    laik_vector_halo(Laik_Instance* inst, Laik_Group* world, Laik_Space* indexSpace, Laik_Partitioning *p1, Laik_Partitioning *p2, Laik_Transition* t1, Laik_Transition* t2, Laik_ReductionOperation operation = LAIK_RO_None);
-    inline T& operator [](int idx) override;
-    T* calc_pointer(int idx, int state);
-    void precalculate_base_pointers() override;
-    void resize(int count) override;
-    void switch_to_p1() override;
-    void switch_to_p2() override;
-    void migrate(Laik_Group* new_group, Laik_Partitioning* p_new_1, Laik_Partitioning* p_new_2, Laik_Transition* t_new_1, Laik_Transition* t_new_2) override;
-};
-
-template <typename T>
-class laik_vector_overlapping:public laik_vector<T>
-{
-public:
-    laik_vector_overlapping(Laik_Instance* inst, Laik_Group* world, Laik_Space* indexSpace, Laik_Partitioning *p1, Laik_Partitioning *p2, Laik_Transition* t1, Laik_Transition* t2, Laik_ReductionOperation operation = LAIK_RO_Sum);
-    inline T& operator [](int idx) override;
-    T* calc_pointer(int idx);
-    void precalculate_base_pointers() override;
-    void resize(int count) override;
-    void switch_to_p1() override;
-    void switch_to_p2() override;
-    void migrate(Laik_Group* new_group, Laik_Partitioning* p_new_1, Laik_Partitioning* p_new_2, Laik_Transition* t_new_1, Laik_Transition* t_new_2) override;
-};
-
-template <typename T>
-class laik_vector_ex_repart:public laik_vector<T>
-{
-public:
-    laik_vector_ex_repart(Laik_Instance* inst, Laik_Group* world, Laik_Space* indexSpace, Laik_Partitioning *p1, Laik_Partitioning *p2, Laik_Transition* t1, Laik_Transition* t2, Laik_ReductionOperation operation = LAIK_RO_None);
-    inline T& operator [](int idx) override;
-    T* calc_pointer(int idx, int state);
-    void precalculate_base_pointers() override;
-    void resize(int count) override;
-    void switch_to_p1() override;
-    void switch_to_p2() override;
-    void migrate(Laik_Group* new_group, Laik_Partitioning* p_new_1, Laik_Partitioning* p_new_2, Laik_Transition* t_new_1, Laik_Transition* t_new_2) override;
-
-private:
-    std::vector<T> data_vector;
-};
-
-template <typename T>
-class laik_vector_overlapping_repart:public laik_vector<T>
-{
-public:
-    laik_vector_overlapping_repart(Laik_Instance* inst, Laik_Group* world, Laik_Space* indexSpace, Laik_Partitioning *p1, Laik_Partitioning *p2, Laik_Transition* t1, Laik_Transition* t2, Laik_ReductionOperation operation = LAIK_RO_None);
-    inline T& operator [](int idx) override;
-    T* calc_pointer(int idx, int state);
-    void precalculate_base_pointers() override;
-    void resize(int count) override;
-    void switch_to_p1() override;
-    void switch_to_p2() override;
-    void migrate(Laik_Group* new_group, Laik_Partitioning* p_new_1, Laik_Partitioning* p_new_2, Laik_Transition* t_new_1, Laik_Transition* t_new_2) override;
-
-private:
-    std::vector<T> data_vector;
-};
-
-template <typename T>
-inline
-T& laik_vector_halo<T>::operator [](int idx){
-    return *(this -> pointer_cache[idx]);
-}
-
-template <typename T>
-inline
-T& laik_vector_ex_repart<T>::operator [](int idx){
-    return this -> data_vector[idx];
-}
-
-template <typename T>
-inline
-T& laik_vector_overlapping<T>::operator [](int idx){
-    return *(this -> pointer_cache[idx]);
-}
-
-template <typename T>
-inline
-T& laik_vector_overlapping_repart<T>::operator [](int idx){
-    return this -> data_vector[idx];
-}
-
 #endif // LAIK_VECTOR
