+++ lulesh.h
@@ -26,8 +26,11 @@
 
 #include <math.h>
 #include <vector>
-//#include <laik_port.h>
 #include <laik_vector.h>
+#include "laik_vector_comm_exclusive_halo.h"
+#include "laik_vector_comm_overlapping_overlapping.h"
+#include "laik_vector_repart_exclusive.h"
+#include "laik_vector_repart_overlapping.h"
 
 //**************************************************
 // Allow flexibility for arithmetic representations
